Integrated the check status deploy logic into the sdk package
